Import and export polygon session state.
Now running *"export smtk model"* will generate a file
(which *"import smtk model"* will read) containing the
internal state data for the polygon session.

Note that for the moment, tessellation information is
included to avoid re-running boost-polygon boolean
operations to load files. This may change in the future
as the cost is pretty small even for large-ish models.

+++ smtk/bridge/polygon/operators/CreateModel.cxx
@@ -112,7 +112,7 @@ smtk::model::OperatorResult CreateModel::operateInternal()
       model.setFloatProperty("normal", smtk::model::FloatList(storage->zAxis(), storage->zAxis() + 3));
       model.setFloatProperty("origin", smtk::model::FloatList(storage->origin(), storage->origin() + 3));
       model.setFloatProperty("feature size", storage->featureSize());
-      model.setIntegerProperty("model scale", storage->modelScale());
+      model.setFloatProperty("model scale", storage->modelScale());
       }
     }
   if (!result)
